feat: improve `main` with better skeleton to run binary
The `main` function now initializes a global logger to receive log output, and
executes a new function named `run` to run the `ocd` binary. The `main` function
catches any `ExitCode` to exit with `ExitCode::Success` status, or print any
error with a backtrace then exit with `ExitCode::Failure` status.

The `run` function is currently just a basic skeleton. It is expected to be
given a set of `OsString` arguments to parse and use for running the `ocd`
binary properly. Logic will be added in the future once core functionality of
the `ocd` binary is coded in.

Signed-off-by: Jason Pena <jasonpena@awkless.com>