feat: add `ConfigPaths`
The `ConfigPaths` object determines expected paths to configuration data. It
implements the `Locator` trait to provide currently expected paths that the
`ocd` binary needs to use for configuration data manipulation in the future.

The `Locator` trait was added to resolve an issue I have been having with the
`directories` crate. I cannot reliably create a failing test case to perform
TDD. The `directories` crate not only uses the HOME environment variable but
uses internal POSIX shell chicanery to determine the user's home directory,
making it _nearly_ impossible to fail in a reliable and cross-platform manner.

Thus, the `Locator` trait exists to make it easier to stub out `ConfigPaths`
when needed for future testing given that I cannot find a way to make
`directories` crate fail. Plus, I do not want to mess with any data in the
user's home directory during testing. The `Locator` trait makes will make it
easier to redirect configuration path data to a temporary directory during
testing.

Signed-off-by: Jason Pena <jasonpena@awkless.com>