feat!: add skeleton of `toml::Toml`
Setup basic skeleton of of Toml parser. Currently, skeleton provides a very
basic CRUD interface that operates on a per table basis. The idea is to
manipulate configuration file data by table and receiving entry data in that
table by key.

Using `toml_edit` crate to take advantage of its format preserving properties.
The plan is to allow the user to modify configuration file data by hand when
they want to, while allowing the application to manipulate that data as needed
without ruining the original formatting the user used.

Signed-off-by: Jason Pena <jasonpena@awkless.com>